+++ lib/rust/parser/src/syntax/operator/types.rs
@@ -1,12 +1,15 @@
-use crate::syntax::operator::operand::Operand;
+use crate::syntax::operator::section::MaybeSection;
 use crate::syntax::token;
 use crate::syntax::tree;
 use crate::syntax::treebuilding::Spacing;
+use crate::syntax::Inspect;
+use crate::syntax::Token;
 use crate::syntax::Tree;
+use crate::syntax::TreeConsumer;
 
-use std::borrow::Cow;
-use std::cmp::Ordering;
-
+use crate::syntax::operator::annotations::Annotation;
+use crate::syntax::operator::named_app::NamedApp;
+use std::fmt::Debug;
 
 
 // ================
@@ -22,26 +25,40 @@ pub struct Operator<'s> {
     pub arity:            Arity<'s>,
 }
 
+impl<'s> Operator<'s> {
+    pub(crate) fn spacing(&self) -> Spacing {
+        match &self.arity {
+            Arity::Unary(token) => Spacing::of_token(token),
+            Arity::Binary { tokens, .. } => Spacing::of_token(tokens.first().unwrap()),
+            Arity::App => Spacing::Spaced,
+            Arity::NamedApp(_) => Spacing::Spaced,
+            Arity::Annotation(annotation) => annotation.spacing(),
+        }
+    }
+}
 
 // === Arity ===
 
 /// Classifies the role of an operator.
 #[derive(Debug)]
 pub enum Arity<'s> {
-    Unary {
-        token: token::Operator<'s>,
-        error: Option<Cow<'static, str>>,
-    },
+    Unary(token::UnaryOperator<'s>),
     Binary {
-        tokens:            Vec<token::Operator<'s>>,
+        tokens:            Vec<Token<'s>>,
         missing:           Option<BinaryOperand>,
         reify_rhs_section: bool,
     },
+    App,
+    NamedApp(Box<NamedApp<'s>>),
+    Annotation(Annotation<'s>),
 }
 
 impl<'s> Arity<'s> {
-    fn unary(token: token::Operator<'s>) -> Self {
-        Self::Unary { token, error: None }
+    pub fn expects_rhs(&self) -> bool {
+        matches!(
+            self,
+            Arity::Unary(_) | Arity::Binary { missing: None | Some(BinaryOperand::Left), .. }
+        )
     }
 }
 
@@ -59,9 +76,8 @@ pub enum BinaryOperand {
 
 #[derive(Debug, Copy, Clone)]
 pub struct ModifiedPrecedence {
-    pub spacing:            Spacing,
-    pub precedence:         token::Precedence,
-    pub is_value_operation: bool,
+    value: u8,
+    mask:  u8,
 }
 
 pub struct ModifiedPrecedenceComparisonResult {
@@ -70,29 +86,27 @@ pub struct ModifiedPrecedenceComparisonResult {
 }
 
 impl ModifiedPrecedence {
+    pub fn new(spacing: Spacing, precedence: token::Precedence, is_value_operation: bool) -> Self {
+        let unspaced_bit = match spacing {
+            Spacing::Spaced => 0,
+            Spacing::Unspaced => 0x80,
+        };
+        let value = precedence.into_u8() | unspaced_bit;
+        let mask = if is_value_operation { 0x7f } else { 0xff };
+        Self { value, mask }
+    }
+
     pub fn compare(&self, other: &Self, include_eq: bool) -> ModifiedPrecedenceComparisonResult {
-        let spacing_ordering = match (self.spacing, other.spacing) {
-            (Spacing::Spaced, Spacing::Unspaced) => Some(Ordering::Less),
-            (Spacing::Unspaced, Spacing::Spaced) => Some(Ordering::Greater),
-            _ => None,
-        };
-        let use_spacing = !(self.is_value_operation && other.is_value_operation);
-        let natural_ordering = self.precedence.cmp(&other.precedence);
-        let natural_is_greater = natural_ordering == Ordering::Greater
-            || (include_eq && natural_ordering == Ordering::Equal);
-        let (is_greater, inconsistent_spacing) = match spacing_ordering {
-            Some(spacing_ordering) => {
-                let spacing_is_greater = spacing_ordering == Ordering::Greater
-                    || (include_eq && spacing_ordering == Ordering::Equal);
-                if use_spacing {
-                    (spacing_is_greater, false)
-                } else {
-                    (natural_is_greater, natural_is_greater != spacing_is_greater)
-                }
-            }
-            None => (natural_is_greater, false),
-        };
+        let adjusted_self = self.value + include_eq as u8;
+        let mask = self.mask | other.mask;
+        let is_greater = adjusted_self & mask > other.value & mask;
+        let is_greater_including_space = adjusted_self > other.value;
+        let inconsistent_spacing = is_greater != is_greater_including_space;
         ModifiedPrecedenceComparisonResult { is_greater, inconsistent_spacing }
+    }
+
+    pub fn min() -> Self {
+        Self { value: 0, mask: 0xff }
     }
 }
 
@@ -129,34 +143,68 @@ impl Warnings {
 // ======================================
 
 pub trait OperandConsumer<'s> {
-    fn push_operand(&mut self, operand: Operand<Tree<'s>>);
+    fn push_operand(&mut self, operand: MaybeSection<Tree<'s>>);
 }
 
 pub trait OperatorConsumer<'s> {
     fn push_operator(&mut self, operator: Operator<'s>);
 }
 
-
-// ===========================
-// === Operator or Operand ===
-// ===========================
-
-#[derive(Debug)]
-pub enum OperatorOrOperand<'s> {
-    Operand(Operand<Tree<'s>>),
-    Operator(Operator<'s>),
+pub trait NamedOperandConsumer<'s> {
+    fn push_maybe_named_operand(&mut self, operand: OperandMaybeNamed<'s>);
 }
 
-impl<'s> From<Operand<Tree<'s>>> for OperatorOrOperand<'s> {
-    fn from(operand: Operand<Tree<'s>>) -> Self {
-        OperatorOrOperand::Operand(operand)
+
+// === Debugging ===
+
+impl<'s, Inner: NamedOperandConsumer<'s>> NamedOperandConsumer<'s> for Inspect<Inner> {
+    fn push_maybe_named_operand(&mut self, operand: OperandMaybeNamed<'s>) {
+        self.observe(&operand);
+        self.0.push_maybe_named_operand(operand);
     }
 }
 
-impl<'s> From<Operator<'s>> for OperatorOrOperand<'s> {
-    fn from(operator: Operator<'s>) -> Self {
-        OperatorOrOperand::Operator(operator)
+impl<'s, Inner: OperatorConsumer<'s>> OperatorConsumer<'s> for Inspect<Inner> {
+    fn push_operator(&mut self, operator: Operator<'s>) {
+        self.observe(&operator);
+        self.0.push_operator(operator);
     }
+}
+
+
+// === Conversions ===
+
+impl<'s, T> OperandConsumer<'s> for T
+where T: NamedOperandConsumer<'s>
+{
+    fn push_operand(&mut self, operand: MaybeSection<Tree<'s>>) {
+        self.push_maybe_named_operand(OperandMaybeNamed::Unnamed(operand));
+    }
+}
+
+impl<'s, T> TreeConsumer<'s> for T
+where T: OperandConsumer<'s>
+{
+    fn push_tree(&mut self, tree: Tree<'s>) {
+        self.push_operand(tree.into());
+    }
+}
+
+
+// ======================
+// === Named Operands ===
+// ======================
+
+#[derive(Debug, PartialEq, Eq)]
+#[allow(clippy::large_enum_variant)] // Clippy considers the `Unnamed` is "at least 0 bytes".
+pub enum OperandMaybeNamed<'s> {
+    Unnamed(MaybeSection<Tree<'s>>),
+    Named {
+        parens:     Option<(token::OpenSymbol<'s>, Option<token::CloseSymbol<'s>>)>,
+        name:       token::Ident<'s>,
+        equals:     token::AssignmentOperator<'s>,
+        expression: Tree<'s>,
+    },
 }